fix: DTO type mismatch, null protection, constant extraction, fallback unification
- PersonalityResultDto: intensity int->float, duration_hours int?->float?
- MoodOffsetCalculator: accept float intensity
- PersonalityThoughtMapper: null protection for thoughts array
- Extract DefaultExcludeKey/MaxTokens/Temperature constants
- shapingHistoryMaxCount fallback: 50->20 (unified with default)
- Fix test assertions for float intensity

+++ Source/Personality/PersonalityResultDto.cs
@@ -14,8 +14,8 @@ namespace RimMind.Personality
         public string type { get; set; } = "state";
         public string label { get; set; } = string.Empty;
         public string description { get; set; } = string.Empty;
-        public int intensity { get; set; }
-        public int? duration_hours { get; set; } = null;
+        public float intensity { get; set; }
+        public float? duration_hours { get; set; } = null;
     }
 
     public class PersonalityIdentityDto
